Refactor #21 : BallFactory 역할과 책임 분리 - BallCreator 생성
