+++ src/main/java/baseball/domain/ball/BallCreator.java
@@ -1,0 +1,9 @@
+package baseball.domain.ball;
+
+import java.util.List;
+
+public interface BallCreator {
+
+    Balls createBalls(List<Integer> numberList);
+
+}
