+++ src/main/java/baseball/domain/ball/BallFactory.java
@@ -3,7 +3,6 @@ package baseball.domain.ball;
 import java.util.List;
 
 public interface BallFactory {
-    List<Integer> numbersToIntegerList(String numbers);
 
-    Balls createBalls(List<Integer> numberList);
+    List<Integer> numbersToIntegerList(String numbers);
 }
