+++ src/main/java/baseball/infrastructure/ball/BallServiceImpl.java
@@ -1,5 +1,6 @@
 package baseball.infrastructure.ball;
 
+import baseball.domain.ball.BallCreator;
 import baseball.domain.ball.BallFactory;
 import baseball.domain.ball.BallService;
 import baseball.domain.ball.Balls;
@@ -9,14 +10,16 @@ import java.util.List;
 public class BallServiceImpl implements BallService {
 
     private final BallFactory ballFactory;
+    private final BallCreator ballCreator;
 
-    public BallServiceImpl(BallFactory ballFactory) {
+    public BallServiceImpl(BallFactory ballFactory, BallCreator ballCreator) {
         this.ballFactory = ballFactory;
+        this.ballCreator = ballCreator;
     }
 
     @Override
     public Balls makeBalls(String numbers) {
         final List<Integer> numberList = ballFactory.numbersToIntegerList(numbers);
-        return ballFactory.createBalls(numberList);
+        return ballCreator.createBalls(numberList);
     }
 }
