+++ src/test/java/baseball/domain/ball/BallCreatorTest.java
@@ -1,0 +1,23 @@
+package baseball.domain.ball;
+
+import baseball.infrastructure.ball.BallCreatorImpl;
+import org.assertj.core.util.Lists;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+import java.util.ArrayList;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+public class BallCreatorTest {
+
+    BallCreator ballCreator = new BallCreatorImpl();
+
+    @Test
+    @DisplayName("Balls 만들기")
+    void createBalls() {
+        final ArrayList<Integer> list = Lists.newArrayList(1, 2, 3);
+        final Balls balls = ballCreator.createBalls(list);
+        assertThat(balls.getBallList().size()).isEqualTo(3);
+    }
+}
